+++ src/spider/worker/TaskExecutor.hpp
@@ -16,8 +16,8 @@
 #include <boost/process/v2/process.hpp>
 #include <boost/process/v2/stdio.hpp>
 
-#include "../utils/BoostAsio.hpp"  // IWYU pragma: keep
-#include "../utils/MsgPack.hpp"  // IWYU pragma: keep
+#include "../io/BoostAsio.hpp"  // IWYU pragma: keep
+#include "../io/MsgPack.hpp"  // IWYU pragma: keep
 #include "FunctionManager.hpp"
 #include "message_pipe.hpp"
