+++ src/spider/worker/message_pipe.cpp
@@ -11,8 +11,8 @@
 #include <fmt/format.h>
 #include <spdlog/spdlog.h>
 
-#include "../utils/BoostAsio.hpp"  // IWYU pragma: keep
-#include "../utils/MsgPack.hpp"  // IWYU pragma: keep
+#include "../io/BoostAsio.hpp"  // IWYU pragma: keep
+#include "../io/MsgPack.hpp"  // IWYU pragma: keep
 
 namespace spider::worker {
